feat: display title in attestation, and remove iframe preview
