+++ app/hypercert/[hypercertId]/components/Attestations/index.tsx
@@ -9,14 +9,20 @@ import UserChip from "@/components/user-chip";
 import { EAS_CONFIGS } from "@/config/eas";
 import { cn } from "@/lib/utils";
 import dayjs from "dayjs";
-import relativeTime from "dayjs/plugin/relativeTime";
-import { Copy, Eye, Globe, Link2, LinkIcon, PlusCircle } from "lucide-react";
+import {
+	ArrowUpRight,
+	Copy,
+	Eye,
+	Globe,
+	Link2,
+	LinkIcon,
+	PlusCircle,
+} from "lucide-react";
 import Link from "next/link";
 import React from "react";
 import SectionWrapper from "../SectionWrapper";
+import Attestation from "./Attestation";
 import URLSourcePreviewDialog from "./URLSourcePreviewDialog";
-
-dayjs.extend(relativeTime);
 
 const Attestations = () => {
 	const hypercert = useFullHypercert();
@@ -42,153 +48,8 @@ const Attestations = () => {
 						</div>
 					) : (
 						validAttestations.slice(0, 2).map((attestation) => {
-							const creationDateFromNow = dayjs(
-								Number(attestation.creationBlockTimestamp) * 1000,
-							).fromNow();
-							const urlSources = attestation.data.sources.filter(
-								(source) => source.type === "url",
-							);
 							return (
-								<div
-									className="flex flex-col gap-4 rounded-lg bg-accent p-4 font-sans"
-									key={attestation.uid}
-								>
-									<div className="flex items-center gap-2">
-										<EthAvatar
-											address={attestation.attester as `0x${string}`}
-											size={36}
-										/>
-										<div className="flex flex-col">
-											<UserChip
-												address={attestation.attester as `0x${string}`}
-												className="border-none p-0 font-bold text-sm"
-												showCopyButton="hover"
-												showAvatar={false}
-											/>
-
-											<span className="text-muted-foreground text-xs">
-												{creationDateFromNow}
-											</span>
-										</div>
-									</div>
-									<p className="">{attestation.data.description}</p>
-									{urlSources.length > 0 && (
-										<div className="flex flex-col gap-1">
-											<span className="font-bold font-sans text-muted-foreground text-xs">
-												Attached Links
-											</span>
-											<div className="flex flex-col gap-1">
-												<URLSourcePreviewDialog
-													urls={urlSources.map((source) => source.src)}
-													TriggerRenderer={({
-														wrapper: Wrapper,
-														setActiveIndex,
-													}) => {
-														return (
-															<>
-																{urlSources
-																	.slice(0, 2)
-																	.map((urlSource, index) => {
-																		return (
-																			<div className="rounded-md border border-border bg-background">
-																				<div
-																					key={`${attestation.uid}-${index}`}
-																					className="flex w-full items-center justify-start gap-2 p-2"
-																				>
-																					<div className="flex h-9 w-9 items-center justify-center rounded-md bg-primary/10">
-																						<LinkIcon
-																							className="text-primary"
-																							size={16}
-																						/>
-																					</div>
-																					<div className="flex flex-1 flex-col gap-1">
-																						<input
-																							type="text"
-																							className="h-auto w-full truncate bg-transparent p-0 text-sm"
-																							value={urlSource.src}
-																							readOnly
-																							disabled
-																						/>
-																						<div className="flex items-center gap-3">
-																							<Link
-																								href={urlSource.src}
-																								target="_blank"
-																							>
-																								<Button
-																									variant={"link"}
-																									className="h-auto gap-1 p-0 text-muted-foreground text-xs"
-																								>
-																									<Globe size={12} />
-																									Visit
-																								</Button>
-																							</Link>
-
-																							<Button
-																								variant={"link"}
-																								className="h-auto gap-1 p-0 text-muted-foreground text-xs"
-																								onClick={() => {
-																									if (window) {
-																										window.navigator.clipboard.writeText(
-																											urlSource.src,
-																										);
-																									}
-																								}}
-																							>
-																								<Copy size={12} />
-																								Copy
-																							</Button>
-																						</div>
-																					</div>
-
-																					<QuickTooltip
-																						content={
-																							<iframe
-																								title={urlSource.src}
-																								src={urlSource.src}
-																								className="overflow-hidden rounded-lg border border-border"
-																							/>
-																						}
-																					>
-																						<Wrapper>
-																							<Button
-																								variant="outline"
-																								size={"sm"}
-																								onClick={() =>
-																									setActiveIndex(index)
-																								}
-																							>
-																								<Eye size={16} />
-																							</Button>
-																						</Wrapper>
-																					</QuickTooltip>
-																				</div>
-																			</div>
-																		);
-																	})}
-																{urlSources.length > 2 && (
-																	<div className="flex items-center justify-center">
-																		<Wrapper>
-																			<Button
-																				variant={"link"}
-																				size={"sm"}
-																				className="text-xs"
-																				onClick={() => {
-																					setActiveIndex(2);
-																				}}
-																			>
-																				View all attachments
-																			</Button>
-																		</Wrapper>
-																	</div>
-																)}
-															</>
-														);
-													}}
-												/>
-											</div>
-										</div>
-									)}
-								</div>
+								<Attestation key={attestation.id} attestation={attestation} />
 							);
 						})
 					)}
